Refactor: move hooks to separate folder
