+++ src/components/Widgets/ContributorsCard/ContributorsCard.tsx
@@ -18,10 +18,10 @@ import { makeStyles } from '@material-ui/core/styles';
 import Alert from '@material-ui/lab/Alert';
 import { InfoCard, Progress } from '@backstage/core';
 import { Entity } from '@backstage/catalog-model';
-import { useProjectEntity } from '../../useProjectEntity';
 import ContributorsList from './components/ContributorsList';
-import { useRequest } from '../../useRequest';
-import { useUrl } from '../../useUrl';
+import { useRequest } from '../../../hooks/useRequest';
+import { useUrl } from '../../../hooks/useUrl';
+import { useProjectEntity } from '../../../hooks/useProjectEntity';
 
 const useStyles = makeStyles(theme => ({
   infoCard: {
