+++ src/components/Widgets/ReadMeCard/ReadMeCard.tsx
@@ -18,9 +18,9 @@ import { makeStyles } from '@material-ui/core';
 import Alert from '@material-ui/lab/Alert';
 import { InfoCard, Progress, MarkdownContent } from '@backstage/core';
 import { Entity } from '@backstage/catalog-model';
-import { useProjectEntity } from '../../useProjectEntity';
-import { useRequest } from '../../useRequest';
-import { useUrl } from '../../useUrl';
+import { useRequest } from '../../../hooks/useRequest';
+import { useUrl } from '../../../hooks/useUrl';
+import { useProjectEntity } from '../../../hooks/useProjectEntity';
 
 const useStyles = makeStyles(theme => ({
   infoCard: {
